+++ Cargo.toml
@@ -7,6 +7,7 @@ members = ["crates/proto",
            "crates/resolver",
            "crates/client", 
            "crates/server",
+           "crates/resolver-async-std",
            "bin",
            "util",
            "tests/compatibility-tests",
