+++ crates/proto/Cargo.toml
@@ -40,7 +40,6 @@ dnssec-openssl = ["dnssec", "openssl"]
 dnssec-ring = ["dnssec", "ring"]
 dnssec = ["data-encoding"]
 testing = []
-async-std-runtime = ["async-std"]
 tokio-runtime = ["tokio/rt-core", "tokio/udp", "tokio/tcp", "tokio/time"]
 default = ["tokio-runtime"]
 
@@ -59,7 +58,6 @@ name = "trust_dns_proto"
 path = "src/lib.rs"
 
 [dependencies]
-async-std = { version = "1.0", optional = true }
 async-trait = "0.1.30"
 backtrace = "0.3.44"
 data-encoding = { version = "2.2.0", optional = true }
