+++ crates/proto/src/udp/udp_stream.rs
@@ -231,22 +231,6 @@ impl<S: UdpSocket> Future for NextRandomUdpSocket<S> {
     }
 }
 
-#[cfg(feature = "async-std-runtime")]
-#[async_trait]
-impl UdpSocket for async_std::net::UdpSocket {
-    async fn bind(addr: &SocketAddr) -> io::Result<Self> {
-        async_std::net::UdpSocket::bind(addr).await
-    }
-
-    async fn recv_from(&mut self, buf: &mut [u8]) -> io::Result<(usize, SocketAddr)> {
-        self.recv_from(buf).await
-    }
-
-    async fn send_to(&mut self, buf: &[u8], target: &SocketAddr) -> io::Result<usize> {
-        self.send_to(buf, target).await
-    }
-}
-
 #[cfg(feature = "tokio-runtime")]
 #[async_trait]
 impl UdpSocket for tokio::net::UdpSocket {
